requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
fastapi-cors
pytube
python-dotenv
ffmpeg
langchain
langchain-community 
langchainhub 
langchain-openai
langchain[docarray]
docarray
whisperx==3.7.2
pytubefix==9.4.1
faiss-cpu==1.12.0